add time_factor calculation to facade weight logic
